Added learning objectives
We are in the process of developing our surveys so that we can have a better understanding of our learners' skills and self-efficacy after taking a Data Carpentry workshop. To do this we want to improve the learning objectives for all of our lessons so that they are measurable. I developed learning objectives with Erin's help for intro to the command line and data wrangling and processing, and Kate advised that the best way to implement these types of changes would be to submit them as a pull request. I would love to have your feedback and input as maintainers of these lessons. This is not my area of expertise, so I'm particularly interested in making sure we have the terminology correct in these LOs.