Include space after the flag '-n' in line 362
Great document! Thanks for coming up with this fantastic resource! 

In line 362, I miss a space after the flag '-n'. Even though without using space the commands works as well, in the previous use of that flag there was always space so I guess to keep the consistency within the document I would add this space.